requirements.txt
flask
wheel
flask_bcrypt
requests
Flask-Cors
stripe
python-dotenv
termcolor
Jinja2
cloudinary
Flask-OAuthlib
Authlib
pymongo
PyJWT
flask_caching
pytz
python-dateutil
itsdangerous
mailtrap
flask-ckeditor[all]
openpyxl
openai